requirements.txt
PySide6>=6.5.0
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
youtube-transcript-api>=0.6.0
openai-whisper>=20231117
pygame>=2.5.2
pytube>=15.0.0
pydub>=0.25.1
lxml>=4.9.3
